backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib==1.7.4
email-validator==2.1.0.post1
gunicorn==21.2.0
pytest==7.4.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23